fix(hal): correct non-address enums and clean up global includes
- Revert mhiti_bits_t and mhitd_bits_t in rx72n_mpu_regs.h from uintptr_t
  to uint32_t; these hold register bitfields and shift amounts, not addresses
- Split rx_flash_memory_addresses_t in rx72n_flash_regs.h: keep uintptr_t for
  address constants, move sizes/blocks/program-units to new rx_flash_memory_layout_t
  (uint32_t)
- Remove redundant src/boot/inc/smc from global include_directories() in
  CMakeLists.txt; it is already covered by target_include_directories()

+++ star-rx72n-firmware/libs/rx_hal/inc/rx72n_mpu_regs.h
@@ -490,7 +490,7 @@ typedef enum : uint16_t {
  * @note Manual ref: Ch17 section 17.2.11
  * @since Version 1.0.0
  */
-typedef enum : uintptr_t {
+typedef enum : uint32_t {
   /** @brief Execute permission in hit region (bit 1) */
   k_mhiti_uhaci_x = 0x00000002U,
 
@@ -560,7 +560,7 @@ typedef enum : uintptr_t {
  * @note Manual ref: Ch17 section 17.2.12
  * @since Version 1.0.0
  */
-typedef enum : uintptr_t {
+typedef enum : uint32_t {
   /** @brief Execute permission in hit region (bit 1) */
   k_mhitd_uhacd_x = 0x00000002U,
